requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
nltk==3.9.2
outcome==1.3.0.post0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
PySocks==1.7.1
python-dotenv==1.2.1
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
selenium==4.36.0
selenium-stealth==1.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
textblob==0.19.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
vaderSentiment==3.3.2
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
